Fixed noramlization bug
Former-commit-id: 437efef938acd63beee9ee8304b5508faaf6c862 [formerly de2d2a194ad5275f624541b9d12fc1c71d46c91d]
Former-commit-id: d6e197f23c5d5e5a89b8e5ba797ba39fd321eb66

+++ normalise.py
@@ -82,7 +82,7 @@ def normalise(wt_paths, mut_paths, outdir, start_indices, end_indices):
         meandiff = meanroi - mean_roi_all
         # subtract the difference from each pixel
         try:
-            imgarr -= meandiff
+            imgarr -= meandiff.astype(np.uint16)  # imagarr = 16bit meandiff = 64bit
         except TypeError:  # Could be caused by imgarr being a short
             imgarr -= int(np.round(meandiff))
